requirements.txt
aiofiles==23.2.1
faiss==1.5.3
faiss_cpu==1.7.3
httpx==0.25.0
langchain==0.0.299
langid==1.1.6
python-dotenv==1.0.0
requests==2.31.0
tiktoken==0.2.0
litellm==0.1.743